% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{lotri}
\alias{.convertId}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lotri}{\code{\link[lotri]{lotri}}}

  \item{rxode2parse}{\code{\link[rxode2parse:dot-convertId]{.convertId}}}
}}

