% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxrandom.R
\name{rxode2randomMd5}
\alias{rxode2randomMd5}
\title{Get the MD5 hash of the current rxode2random revision}
\usage{
rxode2randomMd5()
}
\value{
md5 hash of rxode2random revision
}
\description{
Get the MD5 hash of the current rxode2random revision
}
\examples{
rxode2randomMd5()
}
\author{
Matthew L. Fidler
}
